improvements as of half the period of internship

+++ examples/i1.py
@@ -2,7 +2,7 @@ def computeDeriv(poly):
   new = []
   for i in range(1,len(poly)):
     new.append(
-        float(i*poly[i]))
+        (i*poly[i]))
   if new==[]:
     return 0.0
   return new
